pyproject.toml
[project]
name = "airport-service-api"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.3",
    "django>=6.0.4",
    "django-redis>=6.0.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "psycopg[binary]>=3.3.3",
    "redis>=7.4.0",
    "wait-for-it>=2.3.0",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "celery-types>=0.26.0",
    "django-debug-toolbar>=6.3.0",
    "django-stubs>=6.0.3",
    "djangorestframework-stubs>=3.16.9",
    "mypy>=1.20.2",
    "ruff>=0.15.12",
    "whitenoise>=6.12.0",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "__pycache__",
    "app/airport/migrations/*.py",
    "app/user/migrations/*.py",
]
fix = true
line-length = 120
lint.fixable = ["ALL"]
lint.ignore = [
    "E501", # pycodestyle line too long, handled by black
    "D100", # pydocstyle - missing docstring in public module
    "D101", # pydocstyle - missing docstring in public class
    "D102", # pydocstyle - missing docstring in public method
    "D103", # pydocstyle - missing docstring in public function
    "D104", # pydocstyle - missing docstring in public package
    "D105", # pydocstyle - missing docstring in magic method
    "D106", # pydocstyle - missing docstring in public nested class
    "D107", # pydocstyle - missing docstring in __init__
    "D202", # pydocstyle - no blank lines allowed after function docstring
    "D205", # pydocstyle - 1 blank line required between summary line and description
    "D415", # pydocstyle - first line should end with a period, question mark, or exclamation point
    "UP037", # pyupgrade - removes quotes from type annotation
    "A003", # flake8-builtins - class attribute {name} is shadowing a python builtin
    "B010", # flake8-bugbear - do not call setattr with a constant attribute value
    "B008", # flake8-bugbear - Do not perform function call `Parameter` in argument defaults
    "RUF012", # ruff - mutable class attributes should be annotated with `typing.ClassVar`
    "ANN401", # ruff - Dynamically typed expressions (typing.Any) are disallowed
    "PLR0913", # ruff - Too many arguments to function call
    "PLR2004", # Magic value used in comparison
    "FBT001", # Boolean typed positional argument in function definition
    "FBT002", # Boolean default positional argument in function definition
    "FBT003", # Boolean default positional argument in function definition
    "ARG002", # Unused method argument
    "ARG001", # Unused function argument
    "TD002",
    "TD003",
    "FIX002",
    "PGH003",
    "RUF006",
    "SLF001",
    "PT007",
    "S603",
    "PLW2901", # pylint - for loop variable overwritten by assignment target
    "FBT",
    "PT",
    "TD",
    "PERF203", # ignore for now; investigate
    "RUF029", # function declared async but could be sync
    "CPY001", # add copyright to top of file
    "PLR6301",
    "PLC0415", # import should be at top-level - intentional lazy imports for performance
    "TC002", # move import into type-checking block - conflicts with runtime needs
    "COM812", # missing trailing comma - conflicts with ruff formatter
    "UP047", # pyupgrade - use `X | Y` for union type annotations (py3.10+)
    "ERA001", # eradicate - commented-out code
]
lint.select = ["ALL"]
src = ["app"]
target-version = "py313"

[tool.ruff.format]
preview = true
skip-magic-trailing-comma = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod"]

[tool.ruff.lint.isort]
known-first-party = ["app", "airport", "user", "core"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "D104"]
"app/**/migrations/*.py" = ["D", "E501", "ANN", "UP", "I", "INP001"]
"**/tests/*.py" = ["S101", "PLR2004", "ANN", "SLF001", "ARG001"]

[tool.mypy]
strict = true
ignore_missing_imports = true
mypy_path = "app"
explicit_package_bases = true
